# server/docs/swagger.yaml
definitions:
...
  metrics.Speed:
    properties:
      ID:
        description: 主键ID
        type: integer
      count:
        description: 次数
        type: integer
      createdAt:
        description: 创建时间
        type: string
      exchange:
        description: 交易所
        type: string
      machine_ip:
        description: 机器ip
        type: string
      mean:
        description: 耗时(平均)
        type: number
      metric_name:
        description: 指标名
        type: string
      mid:
        description: 耗时(平均)
        type: number
      sample_time:
        description: 创建时间
        type: string
      std:
        description: 标准差
        type: number
      symbol:
        description: Symbol
        type: string
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
info:
  contact: {}
  description: 使用gin+vue进行极速开发的全栈开发基础平台
  title: Gin-Vue-Admin Swagger API接口文档
  version: v2.6.1
paths:
...
  /metric/createSpeed:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建耗时
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/metrics.Speed'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"创建成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建耗时
      tags:
      - Speed
  /metric/deleteSpeed:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除耗时
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/metrics.Speed'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除耗时
      tags:
      - Speed
  /metric/deleteSpeedByIds:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除耗时
      tags:
      - Speed
  /metric/findSpeed:
    get:
      consumes:
      - application/json
      parameters:
      - description: 主键ID
        in: query
        name: ID
        type: integer
      - description: 次数
        in: query
        name: count
        type: integer
      - description: 创建时间
        in: query
        name: createdAt
        type: string
      - description: 交易所
        in: query
        name: exchange
        type: string
      - description: 机器ip
        in: query
        name: machine_ip
        type: string
      - description: 耗时(平均)
        in: query
        name: mean
        type: number
      - description: 指标名
        in: query
        name: metric_name
        type: string
      - description: 耗时(平均)
        in: query
        name: mid
        type: number
      - description: 创建时间
        in: query
        name: sample_time
        type: string
      - description: 标准差
        in: query
        name: std
        type: number
      - description: Symbol
        in: query
        name: symbol
        type: string
      - description: 更新时间
        in: query
        name: updatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询耗时
      tags:
      - Speed
  /metric/getSpeedList:
    get:
      consumes:
      - application/json
      parameters:
      - in: query
        name: endCreatedAt
        type: string
      - in: query
        name: exchange
        type: string
      - description: 关键字
        in: query
        name: keyword
        type: string
      - in: query
        name: machine_ip
        type: string
      - in: query
        name: metric_name
        type: string
      - in: query
        name: order
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页大小
        in: query
        name: pageSize
        type: integer
      - in: query
        name: sort
        type: string
      - in: query
        name: startCreatedAt
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取耗时列表
      tags:
      - Speed
  /metric/updateSpeed:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新耗时
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/metrics.Speed'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新耗时
      tags:
      - Speed
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"